requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-ninja==1.3.0
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.1
typing_extensions==4.12.2
